[ADD] 4.21 expansion blog list step 9

+++ part4/controllers/blogs.js
@@ -37,7 +37,17 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if(!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if(blog.author.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+  }
+  
   response.status(204).end()
 })
